Methods to get internal libraries
This is consistent with what i did for PakNet. This is better because now dot indexing on the Pack object and any derived objects is only for datatypes. This also promotes localizing these libraries when using them in custom datatypes (see the update to Miscellaneous.md), which is better for performance